Move utility components to a separate package
