+++ csb/src/main/scala/edu/msstate/dasi/csb/util/Util.scala
@@ -1,4 +1,4 @@
-package edu.msstate.dasi.csb
+package edu.msstate.dasi.csb.util
 
 import java.io.File
